added auth and adjust port

+++ start/server.sh
@@ -1,4 +1,8 @@
-for name in ARANGO_DOCKER_NAME ARANGO_PORT ARANGO_MODE ARANGO_STORAGE_ENGINE ARANGO_EDITION; do
+if [ -z "$ARANGO_AUTH" ]; then
+    ARANGO_AUTH="auth"
+fi
+
+for name in ARANGO_DOCKER_NAME ARANGO_PORT ARANGO_MODE ARANGO_STORAGE_ENGINE ARANGO_EDITION ARANGO_AUTH; do
     if [ -z "${!name}" ]; then
         echo "$name missing"
         exit 1
@@ -9,6 +13,7 @@ echo "NAME: $ARANGO_DOCKER_NAME"
 echo "PORT: $ARANGO_PORT"
 echo "MODE: $ARANGO_MODE"
 echo "ENGINE: $ARANGO_STORAGE_ENGINE"
+echo "AUTH: $ARANGO_AUTH"
 echo "EDITION: $ARANGO_EDITION"
 echo
 
@@ -17,29 +22,48 @@ docker rm -fv $ARANGO_DOCKER_NAME > /dev/null 2>&1 || true
 docker pull c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel
 
 if [ "$ARANGO_MODE" == "cluster" ]; then
-    JWTFILE="jwtsecret.$$"
-    rm -f $JWTFILE
+    if [ "ARANGO_AUTH" == "auth" ]; then
+        JWTFILE="jwtsecret.$$"
+        rm -f $JWTFILE
 
-    echo "geheim" > $JWTFILE
+        echo "geheim" > $JWTFILE
 
-    docker run \
-        --name=$ARANGO_DOCKER_NAME \
-        -d \
-        -v $JWTFILE:/jwtsecret:ro \
-        -p $ARANGO_PORT:8529 \
-        -e ARANGO_ROOT_PASSWORD=$ARANGO_ROOT_PASSWORD \
-        c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel \
-        arangodb --starter.local --server.storage-engine $ARANGO_STORAGE_ENGINE --starter.data-dir testrun
+        docker run \
+            --name=$ARANGO_DOCKER_NAME \
+            -d \
+            -v $JWTFILE:/jwtsecret:ro \
+            -p $ARANGO_PORT:8529 \
+            -e ARANGO_ROOT_PASSWORD=$ARANGO_ROOT_PASSWORD \
+            c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel \
+            arangodb --starter.local --server.storage-engine $ARANGO_STORAGE_ENGINE --starter.data-dir testrun
 
-    rm -f $JWTFILE
+        rm -f $JWTFILE
+    else
+        docker run \
+            --name=$ARANGO_DOCKER_NAME \
+            -d \
+            -p $ARANGO_PORT:8529 \
+            c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel \
+            arangodb --starter.local --server.storage-engine $ARANGO_STORAGE_ENGINE --starter.data-dir testrun
+    fi
 elif [ "$ARANGO_MODE" == "singleserver" ]; then
-    docker run \
-        --name=$ARANGO_DOCKER_NAME \
-        -d \
-        -p $ARANGO_PORT:8529 \
-        -e ARANGO_ROOT_PASSWORD=$ARANGO_ROOT_PASSWORD \
-        -e ARANGO_STORAGE_ENGINE=$ARANGO_STORAGE_ENGINE \
-        c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel
+    if [ "ARANGO_AUTH" == "auth" ]; then
+        docker run \
+            --name=$ARANGO_DOCKER_NAME \
+            -d \
+            -p $ARANGO_PORT:8529 \
+            -e ARANGO_ROOT_PASSWORD=$ARANGO_ROOT_PASSWORD \
+            -e ARANGO_STORAGE_ENGINE=$ARANGO_STORAGE_ENGINE \
+            c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel
+    else
+        docker run \
+            --name=$ARANGO_DOCKER_NAME \
+            -d \
+            -p $ARANGO_PORT:8529 \
+            -e ARANGO_NO_PASSWORD=1 \
+            -e ARANGO_STORAGE_ENGINE=$ARANGO_STORAGE_ENGINE \
+            c1.triagens-gmbh.zz:5000/arangodb/linux-${ARANGO_EDITION}-maintainer:devel
+    fi
 else
     echo "unknown mode $ARANGO_MODE"
     exit 1
